# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Masha Api Documentation",
        "description": "Api docs for Masha Application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/missions/lastuserupdate": {
            "get": {
                "tags": [
                    "Missions"
                ],
                "summary": "Get last user update",
                "description": "Retrieves the information about the user who last updated a mission record along with the timestamp of the update.",
                "operationId": "getLastUserUpdate",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "updated_at_date": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "updated_at_time": {
                                                    "type": "string",
                                                    "format": "time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "לא נמצא משתמש."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "התרחש בעיית שרתת יש לנסות שוב מאוחר יותר."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/missions/byyearandmonth": {
            "get": {
                "tags": [
                    "Missions"
                ],
                "summary": "Get missions by year and month",
                "description": "Retrieves missions based on the specified year and month.",
                "operationId": "getMissionsByYearAndMonth",
                "requestBody": {
                    "description": "JSON payload containing the selected year and month",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "selected_month",
                                    "selected_year"
                                ],
                                "properties": {
                                    "selected_month": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "selected_year": {
                                        "type": "integer",
                                        "example": 2023
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "platform": {
                                                "type": "string"
                                            },
                                            "comment": {
                                                "type": "string"
                                            },
                                            "color_comment": {
                                                "type": "integer"
                                            },
                                            "month": {
                                                "type": "integer"
                                            },
                                            "plan_week_per_month": {
                                                "type": "integer"
                                            },
                                            "cumulative_per_month": {
                                                "type": "integer"
                                            },
                                            "year": {
                                                "type": "integer"
                                            },
                                            "plan_week_per_year": {
                                                "type": "integer"
                                            },
                                            "cumulative_per_year": {
                                                "type": "integer"
                                            },
                                            "created_by_user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "employee_type": {
                                                        "type": "integer"
                                                    },
                                                    "permission": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "permission_name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "updated_by_user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "employee_type": {
                                                        "type": "integer"
                                                    },
                                                    "permission": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "permission_name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "חובה לשלוח ערכים של תאריך לחיפוש תקינים."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "התרחש בעיית שרתת יש לנסות שוב מאוחר יותר."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}